backend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==3.1.1
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.5.0
django-extensions==3.0.9
django-nose==1.4.7
django-rest-auth==0.9.5
djangorestframework==3.12.1
drf-yasg2==1.18.5
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
isort==5.5.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
nose==1.3.7
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
